Use u8 for cvar callback IDs instead of u16
Reduces memory footprint of cvar structs and internal callback tables. The callback ID space (MAX_CVAR_CALLBACKS = 128) fits comfortably in a single byte. Also unifies the sentinel value to 0xFF (CVAR_CB_NONE), which works as a byte-fill sentinel in memset() operations and across all u8 fields (function_id, module_id, and free list heads).

+++ source/engine/core/core_api.h
@@ -103,7 +103,7 @@ typedef struct core_api_s
     void        ( *cvar_reset )         ( cvar_t* cv );
 
     /* owner module id is stamped automatically; re-register in reload() */
-    uint16_t    ( *cvar_callback_register )   ( cvar_t* cv, cvar_callback_fn fn );
+    uint8_t     ( *cvar_callback_register )   ( cvar_t* cv, cvar_callback_fn fn );
     void        ( *cvar_callback_unregister ) ( cvar_t* cv );
 
     /* config writer hook: writeconfig appends service sections after cvars + binds */
